requirements.txt
fastapi
uvicorn
transformers
httpx
pydub
python-multipart 
python-dotenv
pymongo
sentencepiece 
accelerate  
langchain
langchain_core
langchain_groq
scipy
streamlit
numpy
requests
audio_recorder_streamlit
torchaudio
